requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
rembg==2.0.57
Pillow==10.4.0
python-multipart==0.0.12
google-genai==0.8.3
